requirements.txt
fastapi[all]
tensorflow
numpy
Pillow
uvicorn